# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "BFD",
    "contact": {}
  },
  "host": "localhost:8080",
  "basePath": "/api",
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
    }
  },
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/user/onboarding": {
      "get": {
        "summary": "Onboarding",
        "tags": [
          "User"
        ],
        "operationId": "Onboarding",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/organizer/get": {
      "post": {
        "summary": "Get Organizers",
        "tags": [
          "Organizer"
        ],
        "operationId": "GetOrganizers",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/GetOrganizersRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/organizer": {
      "get": {
        "summary": "Get Organizer",
        "tags": [
          "Organizer"
        ],
        "operationId": "GetOrganizer",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/organizer/update": {
      "post": {
        "summary": "Update Organizer",
        "tags": [
          "Organizer"
        ],
        "operationId": "UpdateOrganizer",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "firstName",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "lastName",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "company",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "email",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "phone",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "userType",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "organizer",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "package",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "designation",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "employeeType",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "holiday",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "userLimit",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "workType",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "weeklyWorkHours",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "isImmutable",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "profilePicture",
            "in": "formData",
            "required": true,
            "type": "file",
            "format": "file",
            "description": ""
          },
          {
            "name": "address[streetNo]",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "address[city]",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "address[zip]",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "id",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "GetOrganizersRequest": {
      "title": "GetOrganizersRequest",
      "example": {
        "sortBy": "",
        "order": "",
        "search": "",
        "page": 1,
        "limit": 10,
        "organizer": ""
      },
      "type": "object",
      "properties": {
        "sortBy": {
          "type": "string"
        },
        "order": {
          "type": "string"
        },
        "search": {
          "type": "string"
        },
        "page": {
          "type": "integer",
          "format": "int32"
        },
        "limit": {
          "type": "integer",
          "format": "int32"
        },
        "organizer": {
          "type": "string"
        }
      },
      "required": [
        "sortBy",
        "order",
        "search",
        "page",
        "limit",
        "organizer"
      ]
    },
...
  },
...
}